can't get mongoose working
I typed ‘mongo’ in one terminal and ‘mongo’ in another and it seems to
work.

Then I tried this:
‘mongo trucks < models/foodTrucksJSON.js’

And it seemed to work because I got this as part of the message:
"nInserted" : 60,

When I try this, there’s nothing returned. Can you tell me what I might
be doing wrong?
‘db.trucks.find()’